Move _topology module out of dimensions subpackage.
This allows a dependency order of

    core._topology -> core.timespan -> core.dimensions

which will allow DatabaseTimespanRepresentation to depend on
TopologicalExtentDatabaseRepresentation, as intended.
