+++ python/lsst/daf/butler/core/dimensions/_skypix.py
@@ -39,9 +39,9 @@ import sqlalchemy
 from lsst.sphgeom import Pixelization
 from lsst.utils import doImport
 from .. import ddl
+from .._topology import TopologicalFamily, TopologicalRelationshipEndpoint, TopologicalSpace
 
 from ..named import NamedValueAbstractSet, NamedValueSet
-from ._topology import TopologicalFamily, TopologicalRelationshipEndpoint, TopologicalSpace
 from ._elements import Dimension
 from .construction import DimensionConstructionBuilder, DimensionConstructionVisitor
